Share test utilities (`like` / `wait` and error assertions) between unit/integration and consumer tests
