+++ test/__fixtures__/index.ts
@@ -1,0 +1,4 @@
+
+export * from './expectations';
+export * from './like';
+export * from './wait';
