+++ test/unit/__fixtures__/child-processes.ts
@@ -1,5 +1,5 @@
 import { MockChildProcess, mockChildProcessModule } from '../__mocks__/mock-child-process';
-import { wait } from './wait';
+import { wait } from '../../__fixtures__';
 
 const EXIT_CODE_SUCCESS = 0;
 const EXIT_CODE_ERROR = 1;
