+++ test/unit/__fixtures__/index.ts
@@ -2,9 +2,11 @@
 export * from './child-processes';
 export * from './expectations';
 export * from './instance';
-export * from './like'
-export * from './wait'
 
 export * from './responses/diff';
 export * from './responses/merge';
 export * from './responses/remote-messages';
+
+export * from '../../__fixtures__/expectations'
+export * from '../../__fixtures__/like'
+export * from '../../__fixtures__/wait'
